test: Refactor integration fixture tests into reusable function
